# only reset sqlite3 statement when necessary